Upgrade jflex to 1.8.2
 patch by Mick Semb Wever; reviewed by Brandon Williams for CASSANDRA-16576

+++ src/java/org/apache/cassandra/index/sasi/analyzer/StandardTokenizerInterface.java
@@ -35,7 +35,7 @@ public interface StandardTokenizerInterface
     /**
      * Returns the current position.
      */
-    int yychar();
+    long yychar();
 
     /**
      * Returns the length of the matched text region.
